Jquery to add decoration min quantity and to cart

+++ db/migrate/20170924164250_add_minimum_size_to_decorations.rb
@@ -1,5 +1,5 @@
 class AddMinimumSizeToDecorations < ActiveRecord::Migration[5.1]
   def change
-    add_column :decorations, :minimum_size, :integer, default: 0, null: false
+    add_column :decorations, :minimum_size, :float, default: 0.0, null: false
   end
 end
